Refactor App component
